V1/requirements.txt
pyserial
numpy
matplotlib
sumo
traci
